Update dali2mqtt.service
Change to make use that network on up and online as it will otherwise fail because the mqtt server is not reachable. 

When looking into this i noted that a better way of solving this is to have the python not crash when the server is not reached and just keep on trying to connect. This should be doable with connect_async instead of connect but it needs some other changes aswell.